データベーススキーマの変更とエンティティのフィールド名を更新し、テーブル名をプレフィックス付きに変更。ユーザーIDの型をLongからIntに変更し、関連するリポジトリとサービスを修正。新しいマイグレーションファイルを追加。

+++ src/main/kotlin/com/example/myapp/dto/LoginResponse.kt
@@ -4,7 +4,7 @@ data class LoginResponse(
     val success: Boolean,
     val message: String,
     val requireTotp: Boolean = false,
-    val userId: Long? = null,
+    val userId: Int? = null,
     val username: String? = null,
     val email: String? = null
 )
