+++ src/main/kotlin/com/example/myapp/dto/RegisterResponse.kt
@@ -3,6 +3,6 @@ package com.example.myapp.dto
 data class RegisterResponse(
     val success: Boolean,
     val message: String,
-    val userId: Long? = null,
+    val userId: Int? = null,
     val username: String? = null
 )
